Configurazione e implementazione Spring Security

+++ src/main/webapp/WEB-INF/views/esami/tipologie.jsp
@@ -1,6 +1,7 @@
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@taglib prefix="form" uri="http://www.springframework.org/tags/form"%>
+<%@ taglib prefix="sec" uri="http://www.springframework.org/security/tags" %>
 
 <style>
     .chip {
@@ -18,33 +19,37 @@
 <div class="panel panel-default">
     <div class="panel-heading">
         <div class="panel-title pull-left">Tipologie Esami
-            &ensp;
-            <i class="fa fa-plus-circle" aria-hidden="true"></i>
-            <a align="right" href="#" data-toggle="modal" data-target="#indicatoreModal">
-                Aggiungi indicatore
-            </a>
-            &ensp;
-            <i class="fa fa-plus-circle" aria-hidden="true"></i>
-            <a align="right" href="#" data-toggle="modal" data-target="#prerequisitoModal">
-                Aggiungi prerequisito
-            </a>
+            <sec:authorize access="hasRole('ROLE_ADMIN')">
+                &ensp;
+                <i class="fa fa-plus-circle" aria-hidden="true"></i>
+                <a align="right" href="#" data-toggle="modal" data-target="#indicatoreModal">
+                    Aggiungi indicatore
+                </a>
+                &ensp;
+                <i class="fa fa-plus-circle" aria-hidden="true"></i>
+                <a align="right" href="#" data-toggle="modal" data-target="#prerequisitoModal">
+                    Aggiungi prerequisito
+                </a>
+            </sec:authorize>
         </div>
-        <div class="panel-title pull-right">
-            <i class="fa fa-plus" aria-hidden="true"></i>
-            <a align="right" href="#" data-toggle="modal" data-target="#tipologiaModal">
-                Aggiungi nuova tipologia
-            </a>
-        </div>
+        <sec:authorize access="hasRole('ROLE_ADMIN')">
+            <div class="panel-title pull-right">
+                <i class="fa fa-plus" aria-hidden="true"></i>
+                <a align="right" href="#" data-toggle="modal" data-target="#tipologiaModal">
+                    Aggiungi nuova tipologia
+                </a>
+            </div>
+        </sec:authorize>
         <div class="clearfix"></div>
     </div>
     <div class="panel-body">
         <table id="tipologieTable" class="table table-striped" width="100%">
             <thead>
                 <tr>
-                    <th></th>
+                    <th>Dettagli</th>
                     <th>Nome</th>
                     <th>Codice</th>
-                    <th>Descrizione</th>
+                    <th>Descrizione:</th>
                     <th>Costo</th>
                     <th>Indicatori:</th>
                     <th>Prerequisiti:</th>
@@ -71,7 +76,7 @@
                             </c:forEach>
                         </td>
                         <td>
-                            <a href="${pageContext.request.contextPath}/esami/tipologie/delete?id=${tipologia.id}">
+                            <a class="confirmation" href="${pageContext.request.contextPath}/esami/tipologie/delete?id=${tipologia.id}">
                                 <span class="glyphicon glyphicon-remove"></span>
                             </a>
                         </td>
@@ -91,29 +96,39 @@
                 <h4 class="modal-title">Inserisci dati nuova tipologia</h4>
             </div>
             <div class="modal-body">
-                <form:form method="POST" action="${pageContext.request.contextPath}/esami/tipologie/save" commandName="tipologiaBean">
-                    <div class="form-group-lg">
+                <form:form data-toggle="validator" method="POST" action="${pageContext.request.contextPath}/esami/tipologie/save" commandName="tipologiaBean">
+                    <div class="form-group has-feedback">
                         <div class="input-group">
                             <div class="input-group-addon"><i class="fa fa-file-text fa-lg"></i></div>
-                                <form:input path="nome" id="nome" type="text" class="form-control" placeholder="Nome"/>
+                                <form:input path="nome" id="nome" type="text" class="form-control" placeholder="Nome" required="true"/>
                         </div>
-                        <br>
+                        <span class="glyphicon form-control-feedback" aria-hidden="true"></span>
+                        <div class="help-block with-errors"></div>
+                    </div>
+                    <div class="form-group has-feedback">
                         <div class="input-group">
                             <div class="input-group-addon"><i class="fa fa-bars fa-lg"></i></div>
-                                <form:input path="codice" id="codice" type="text" class="form-control" placeholder="Codice"/>
+                                <form:input path="codice" id="codice" type="text" class="form-control" placeholder="Codice" required="true"/>
                         </div>
-                        <br>
+                        <span class="glyphicon form-control-feedback" aria-hidden="true"></span>
+                        <div class="help-block with-errors"></div>
+                    </div>
+                    <div class="form-group has-feedback">
                         <div class="input-group">
                             <div class="input-group-addon"><i class="fa fa-list fa-lg"></i></div>
-                                <form:input path="descrizione" id="descrizione" type="text" class="form-control" placeholder="Descrizione"/>
+                                <form:input path="descrizione" id="descrizione" type="text" class="form-control" placeholder="Descrizione" required="true"/>
                         </div>
-                        <br>
+                        <span class="glyphicon form-control-feedback" aria-hidden="true"></span>
+                        <div class="help-block with-errors"></div>
+                    </div>
+                    <div class="form-group has-feedback">
                         <div class="input-group">
                             <div class="input-group-addon"><i class="fa fa-eur fa-lg"></i></div>
-                                <form:input path="costo" id="costo" type="text" class="form-control" placeholder="Costo"/>
+                                <form:input path="costo" id="costo" type="text" class="form-control" placeholder="Costo" required="true"/>
                         </div>
+                        <span class="glyphicon form-control-feedback" aria-hidden="true"></span>
+                        <div class="help-block with-errors"></div>
                     </div>
-                    <br>
                     <button type="submit" class="btn btn-success">Salva</button>
                 </form:form>
             </div>
@@ -137,11 +152,12 @@
                         <span class="glyphicon form-control-feedback" aria-hidden="true"></span>
                         <div class="help-block with-errors"></div>
                     </div>
-                    <div class="form-group">
+                    <div class="form-group has-feedback">
                         <label for="selectTipologia">Scegli a quale tipologia associare questo indicatore:</label>
                         <form:select id="selectTipologia" path="tipologiaEsame.id" class="form-control" style="width: 75%" required="true">
                             <form:options items="${listaTipologie}" itemLabel="nome" itemValue="id"></form:options>
                         </form:select>
+                        <div class="help-block with-errors"></div>
                     </div>
                     <button type="submit" class="btn btn-success">Salva</button>
                 </form:form>
@@ -186,6 +202,11 @@
 
 <script>
     $(document).ready(function () {
+        $('.confirmation').on('click', function () {
+            return confirm('ATTENZIONE!\n'+
+                'Eliminando questa tipologia verranno eliminati anche tutti gli esami associati.\n'+
+                'Sei sicuro?');
+        });
         $('#tipologieTable').dataTable({
             "order": [],
             "responsive": true,
